Web/Pope/src/requirements.txt
Flask==2.0.3
PyJWT==2.3.0
requests==2.26.0
Werkzeug==2.2.2

